requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
dj-database-url==2.2.0
Django==4.2.14
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
psycopg==3.2.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
